Refactor database initialization and add schema.sql
This commit refactors the database initialization code in `main.rs` to read the SQL schema from a file (`schema.sql`) and execute it using `sqlx::query`. It also adds the `schema.sql` file to the project, which contains the SQL statements for creating the `LOCATION_TYPES`, `LOCATIONS`, and `LABWARES` tables.

The changes include:
- Importing the `std::fs` module to read the `schema.sql` file
- Using `fs::read_to_string` to read the contents of the file
- Executing the SQL statements from `schema.sql` using `sqlx::query`
- Adding the `LOCATION_TYPES` table creation statement to `schema.sql`

This refactor improves code organization and makes it easier to manage the database schema.